recommender_cosine
i fist tried to use cosin but due to large datast it required a lagre ram space so i then continue with knn


This notebook represents the first attempt at building a content-based recommender using sklearn.metrics.pairwise.cosine_similarity.This approach was not successful and has been superseded by the KNN model.The code is being committed as a record of the development process. The key takeaway was discovering that this method does not scale to our 80,000+ song dataset.Problem: The attempt to build the full similarity matrix ($81k \times 81k$) resulted in a RAM crash, as it required over 50GB of memory. This demonstrates the $O(n^2)$ scalability issue with this method.